Potion Effect rework
All objects that can have potion effects applied to them now need a component on them that inherits from PotionTargetAbstract. What a potion does to an object is defined in this component's OnPotionEffectsAdded and OnPotionEffectsRemoved methods. It is an abstract class so we can define what potion effects should do when applied to an item by making a new class that inherits from PotionTargetAbstract and overriding the OnPotionEffectsAdded/Removed methods. This is mostly for weird cases, like hands, which will probably need to react to potions differently than most normal items. Any generic objects will need the PotionTargetDefault component for potions to work on them. I haven't added any new potion effects so we still only have shrink. I haven't touched anything related to PotionObject. I also haven't touched anything related to the potion vfx so that might need to be looked into

The potion effect scriptable object now has an enum for a conflicting effect. Attempting to add a potion effect while an object has the conflicting effect (Adding shrink to an object with grow) will remove the conflicting effect before the new one is applied