Mallet and stuff
Added the mallet. You can try it in my sandbox. You can pick it up and hit the red ingredient (the red-ish sphere) with it to change it into a different object. This will be used for smashing the red berry ingredient which you can then add to a bottle to make the grow potion

The mallet hitting things and changing them is done with the new RecipeObjectConversion scriptable object. ObjectRoots can have an ingredient name. If the mallet hits something with a specific ingredient name, it will run the scriptable object's ConvertObjectToOtherObject method. This deletes the input object and spawns a set prefab in its place. The new object will also copy any potion effects the previous object had if possible.

The XR Grab Interactable has a setting called Track Scale that's turned on by default. It's meant to make the scale of objects match any changes in scale of your hands, but of course, that conflicts with our size changing potions. Any grabbable objects that can have potion effects on them should disable the Track Scale option