ObjectRootTrigger done
ObjectRootTrigger is now done. It is a trigger volume that detects when objects with certain potion effects and/or objects of a certain kind enter or exit. It has events you can hook up in editor so other people should be able to use it pretty easily. This'll be used for puzzles and whatever

Added reference to ObjectRoot in PotionTargetAbstract. I have often been told GetComponent is pretty bad on performance, so I like saving these references to components whenever it's convenient so I don't have to GetComponent very much.

Added a new OnPotionEffectsChangedEvent that can be subscribed to in code for whatever