Potion shader
The potion shader is now in. There is only one material for it which is on the potion liquid object by default. The colors ("_MainColor" and "_MixedColor") are changed in code so we don't need a separate material for every possible combination of potion. While the bottle is empty, the liquid renderer is disabled

Since the liquid only uses one material, the Material of the PotionEffects has been replaced with a Color

I haven't made a visual for inverted effects yet, haven't decided what I'm going to do with that yet

Messed with PotionObject a bit. I changed the currentEffect to use the enum instead of a scriptable object so it supports multiple effects.
I made the potion bottle default to 1,1,1 scale and I fixed the import settings on the bottle so it's still roughly the same size as it was. Alan, if you want to get your particles to scale with objects correctly, you'll probably have to fix this for a lot of the other objects that imported really tiny. Set the scaling mode on the particle systems to "Hierarchy" if you haven't already to make it match the scale of whatever object it is parented to