Imported assets, mixing functional
Imported assets from the first two asset packs posted in the Discord. They are located in the Third Party Assets folder

Made a solution to the stupid enum/scriptable-object mess me and Alan have been dealing with. Now, you can do enum.GetScriptableObject() and it will return the scriptable object that matches the enum. A new object needs to be placed in the scene for this to work. ALL SCENES MUST HAVE A POTIONCATALOG OBJECT SOMEWHERE IN IT! This is a prefab in the Prefabs > Potions folder. Just put in anywhere in the scene.

Dropping multiple ingredients into a potion will now replace the effect instead of adding to it. You need to use the mixer to make a potion with multiple effects. I haven't tested the mixer much but it should work

To use the mixer, pour two different kinds of potions in it. I don't think I remembered to set a limit of max having two potions mixed together, so I'll have to do that later