Small fixes
The AddPotion/RemovePotion bug Alan messaged me about was because removing conflicting potion effects was only on the ScriptableObject overload of the add method and not the enum overload. The ScriptableObject overload was never used, so the removing of conflicting potion effects never ran. The ScriptableObject overload was useless anyway now that .GetScriptableObject() exists so I got rid of it and added the removal of conflicting effects to the enum AddPotionEffect

There were issues with the VFX object not being added soon enough to apply vfx to objects that were just created through mallet smashing. I moved the vfx object creation into its own method so I can run it right before potion effects get added

Slightly changed the colors of inverted potion effects so there's a visual difference. I was gonna do something special with the shader but I don't care anymore