Menu stuff added, but not finished.
I've added some stuff like the buttons and even different logos that I will be working on tomorrow. I am still struggling a lot with the book ui stuff. Mostly with the buttons and getting them to work properly with to much of a struggle. So I think I will need help with that please.