Added cleanser. Animated droplet
Added cleansing bottle. This is an infinitely reusable white potion that removes the potion effects of any object it hits. I made the vfx for it too

The default potions are now shrink, light, and cold. The other ones must be gotten via inverting

I made the potion droplets have a bit of squash and stretch based on their velocity